remove fhir controller
