+++ src/main/java/org/openelisglobal/fhir/providers/PractitionerProvider.java
@@ -1,25 +1,38 @@
 package org.openelisglobal.fhir.providers;
 
+import ca.uhn.fhir.model.api.Include;
 import ca.uhn.fhir.rest.annotation.Create;
 import ca.uhn.fhir.rest.annotation.Delete;
 import ca.uhn.fhir.rest.annotation.IdParam;
+import ca.uhn.fhir.rest.annotation.IncludeParam;
+import ca.uhn.fhir.rest.annotation.OptionalParam;
 import ca.uhn.fhir.rest.annotation.Read;
 import ca.uhn.fhir.rest.annotation.ResourceParam;
+import ca.uhn.fhir.rest.annotation.Search;
 import ca.uhn.fhir.rest.annotation.Update;
 import ca.uhn.fhir.rest.api.MethodOutcome;
+import ca.uhn.fhir.rest.param.DateRangeParam;
+import ca.uhn.fhir.rest.param.StringAndListParam;
+import ca.uhn.fhir.rest.param.TokenAndListParam;
 import ca.uhn.fhir.rest.server.IResourceProvider;
 import ca.uhn.fhir.rest.server.exceptions.InternalErrorException;
 import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
 import ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException;
 import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException;
 import jakarta.servlet.http.HttpServletRequest;
+
+import java.util.HashSet;
 import java.util.UUID;
 import org.hl7.fhir.instance.model.api.IBaseResource;
+import org.hl7.fhir.r4.model.Bundle;
+import org.hl7.fhir.r4.model.Encounter;
 import org.hl7.fhir.r4.model.IdType;
 import org.hl7.fhir.r4.model.OperationOutcome;
 import org.hl7.fhir.r4.model.Practitioner;
+import org.hl7.fhir.r4.model.ServiceRequest;
 import org.openelisglobal.common.log.LogEvent;
 import org.openelisglobal.common.util.ControllerUtills;
+import org.openelisglobal.dataexchange.fhir.FhirUtil;
 import org.openelisglobal.dataexchange.fhir.exception.FhirLocalPersistingException;
 import org.openelisglobal.dataexchange.fhir.service.FhirPersistanceService;
 import org.openelisglobal.dataexchange.fhir.service.FhirTransformService;
@@ -27,6 +40,7 @@ import org.openelisglobal.person.service.PersonService;
 import org.openelisglobal.person.valueholder.Person;
 import org.openelisglobal.provider.service.ProviderService;
 import org.openelisglobal.provider.valueholder.Provider;
+import org.openelisglobal.spring.util.SpringContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -242,6 +256,33 @@ public class PractitionerProvider implements IResourceProvider {
             LogEvent.logError(this.getClass().getSimpleName(), method,
                     "Unexpected error while deleting Practitioner: " + e.getMessage());
             throw new InternalErrorException("Unexpected server error while deleting Practitioner", e);
+        }
+    }
+
+    @Search
+    public Bundle searchPractitionerBundle(
+               @OptionalParam(name = Practitioner.SP_IDENTIFIER) TokenAndListParam identifier,
+	        @OptionalParam(name = Practitioner.SP_GIVEN) StringAndListParam given,
+	        @OptionalParam(name = Practitioner.SP_FAMILY) StringAndListParam family,
+	        @OptionalParam(name = Practitioner.SP_RES_ID) TokenAndListParam id,
+	        @OptionalParam(name = "_lastUpdated") DateRangeParam lastUpdated,
+	        @IncludeParam(reverse = true, allow = { "Encounter:" + Encounter.SP_PARTICIPANT, "ServiceRequest:" + ServiceRequest.SP_REQUESTER,
+	                 }) HashSet<Include> revIncludes,HttpServletRequest request) {
+        
+        String methodName = "searchPractitionerBundle";
+        LogEvent.logDebug(this.getClass().getSimpleName(), methodName, "Searching for Practitioners (returning Bundle)");
+        
+        try {
+            FhirUtil util = SpringContext.getBean(FhirUtil.class);
+            
+            Bundle bundle = util.forwardSearchToFhirStore(request);
+            
+            return bundle;
+            
+        } catch (Exception e) {
+            LogEvent.logError(this.getClass().getSimpleName(), methodName,
+                "Error searching Practitioners: " + e.getMessage());
+            throw new InternalErrorException("Error searching Practitioners", e);
         }
     }
 }
